WIP: Seperate components from main file.
