+++ app/data/versionInfo.ts
@@ -1,0 +1,14 @@
+const versionInfo = {
+    version: "1.5.0 Beta",
+    released: "July 28, 2025",
+    platforms: {
+        windows: {
+            url: "https://apps.microsoft.com/detail/9PBWHHZRQFW6?hl=en-us&gl=US&ocid=pdpshare",
+            size: "31 MB",
+        },
+        mac: { url: "https://example.com/mac", size: "123 MB" },
+        linux: { url: "https://example.com/linux", size: "123 MB" },
+    },
+};
+
+export default versionInfo;
