+++ app/layout.tsx
@@ -1,34 +1,35 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
+import Footer from "./components/Footer/Footer";
+import Header from "./components/Header/Header";
 
 const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
+    variable: "--font-geist-sans",
+    subsets: ["latin"],
 });
 
 const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
+    variable: "--font-geist-mono",
+    subsets: ["latin"],
 });
 
 export const metadata: Metadata = {
-  title: "Icarus - Valorant Strategy Planner",
-  description: "Revolutionary Valorant strategy planning tool",
+    title: "Icarus - Valorant Strategy Planner",
+    description: "Revolutionary Valorant strategy planning tool",
+    icons: "./favicon.ico",
 };
 
 export default function RootLayout({
-  children,
+    children,
 }: Readonly<{
-  children: React.ReactNode;
+    children: React.ReactNode;
 }>) {
-  return (
-    <html lang="en">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
-        {children}
-      </body>
-    </html>
-  );
+    return (
+        <html lang="en">
+            <Header />
+            <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>{children}</body>
+            <Footer />
+        </html>
+    );
 }
