+++ app/types/AnimationVariants.ts
@@ -1,0 +1,7 @@
+import { Variants } from "framer-motion";
+
+export type AnimationVariants = {
+    fadeUp: Variants;
+    fadeIn: Variants;
+    pop: Variants;
+};
